Use provided buffer size when serving non-Path Resources
The helper IOResources.asContentSource wraps a given Resource to be
served as a Content.Source. Unless a resource had a Path component,
InputStreamContentSource is used as a fallback.

Previously, we were using a hardcoded buffer size of 4096, even if the
request size was larger (e.g., 32768).

This severely limited throughput when serving content via
non-Path-bearing Resource implementations.

Use the provided buffer pool and size instead.

Fixes #12963

Signed-off-by: Christian Kohlschütter <christian@kohlschutter.com>