Fixes #13164 - ProxyServlet abort. (#13169)
* Fixed HttpStatus javadocs and added isError(int).
* Fixed double invocations of sendProxyResponseError().
For ProxyServlet and AsyncProxyServlet, the proxyRequest is always sent early, so ProxyResponseListener.onComplete() will take care of calling sendProxyResponseError(), so it is enough to cause the abortion of the proxyRequest.
* Fixed double invocations of onClientRequestFailure(), which was aborting the proxyRequest, but was calling sendProxyResponseError() if the abort was not successful.
Aborting the proxyRequest was already calling sendProxyResponseError(), so no need to calling it again if the abort was not successful.
* AsyncMiddleManServlet is different, since it does not send the proxyRequest immediately.
Therefore added an attribute to know whether it was sent, and acting accordingly to guarantee that sendProxyResponseError() is called even if the proxyRequest is not sent yet (since there is no ProxyResponseListener.onComplete() installed yet).

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>
Co-authored-by: Olivier Lamy <olamy@apache.org>